backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.17.0
attrs==25.4.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
eventlet==0.40.3
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
mac-vendor-lookup==0.1.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
netifaces==0.11.0
pillow==11.3.0
propcache==0.4.1
psutil==7.1.0
PyJWT==2.10.1
python-engineio==4.12.3
python-nmap==0.7.1
python-socketio==5.14.1
reportlab==4.4.4
requests==2.32.5
scapy==2.6.1
simple-websocket==1.1.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
